+++ lib/src/puppeteer.dart
@@ -292,7 +292,8 @@ class Puppeteer {
 
     return [
       ..._defaultArgs,
-      if (userDataDir != null) '--user-data-dir=$userDataDir',
+      if (userDataDir != null)
+        '--user-data-dir=${p.join(Directory.current.path, userDataDir)}',
       if (noSandboxFlag) '--no-sandbox',
       if (devTools) '--auto-open-devtools-for-tabs',
       if (headless) ..._headlessArgs,